% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens.R
\name{get_dens}
\alias{get_dens}
\title{Get the estimated density in each knot and make dataframe}
\usage{
get_dens(category_name)
}
\arguments{
\item{category_name}{change numeric category to category name}
}
\description{
Get the estimated density in each knot from Save.RData.
}
