% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cog.R
\name{map_cog}
\alias{map_cog}
\title{Plot the COG on map from VAST outputs}
\usage{
map_cog(
  pcategory_name,
  zone,
  labs,
  ncol,
  shape,
  size,
  DateFile,
  package,
  fileEncoding
)
}
\arguments{
\item{zone}{no. of UTM which was estimated when "Extrapolation_List" was made}

\item{labs}{setting of figure. colour is legend}

\item{ncol}{number of figures in line side by side. max is no. of "Category"}

\item{shape}{shape of COG point}

\item{size}{size of shape}

\item{DateFile}{necessary to make a COG_Table using VAST}

\item{package}{Spatial DeltaGLMM or FishStatsUtils}

\item{fileEncoding}{endoding}

\item{category_name}{names of each category. set 1 when single-species model}
}
\description{
Plot the COG on map from VAST outputs
}
\examples{
map_cog(pcategory_name, zone, labs, ncol, shape, size, DateFile, package, fileEncoding)
}
