% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dens.R
\name{map_dens}
\alias{map_dens}
\title{Plot the estimated density on map}
\usage{
map_dens(data, region, scale_name, ncol, shape, size)
}
\arguments{
\item{data}{dataframe made by get_dens function}

\item{region}{region of the data.}

\item{ncol}{number of figures in line side by side. max is no. of "Category"}

\item{shape}{shape of COG point}

\item{size}{size of shape}

\item{scale_names}{unit of estimated density}
}
\description{
plot the estimated density om map by species using ggplot2. scale name, plot shape, and size of plot shape can change.
}
