% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Get the estimated density in each knot and make dataframe}
\usage{
plot_index(vast_index, DG, category_name, fig_output_dirname)
}
\arguments{
\item{vast_index}{Standardized data}

\item{DG}{nominal data}

\item{category_name}{category name}

\item{fig_output_dirname}{directory for output figure}
}
\description{
Get the estimated density in each knot from Save.RData.
}
