% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cog.R
\name{map_cog}
\alias{map_cog}
\title{Plot the COG on map from VAST output}
\usage{
map_cog(
  data_type,
  category_name,
  region,
  ncol,
  shape,
  size,
  zoom_in_lon,
  zoom_in_lat,
  fig_output_dirname
)
}
\arguments{
\item{data_type}{VAST ot nomical}

\item{category_name}{names of each category}

\item{region}{region of the data. please see require(maps) and unique(map_data("world")$region)}

\item{ncol}{number of figures in line side by side. max is no. of "Category"}

\item{shape}{shape of COG point}

\item{size}{size of shape}

\item{zoom_in_lon}{zoom in on the map if 1<, zoom out on the map if 1>, and 1 is the same size}

\item{zoom_in_lat}{zoom in on the map if 1<, zoom out on the map if 1>, and 1 is the same size}

\item{fig_output_dirname}{directory for output}
}
\description{
plot the COG om map using ggplot2. category names, axis names, font size and plot shape can change.
}
