% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cog.R
\name{\%>\%}
\alias{\%>\%}
\title{Plot the COG on map from VAST output}
\arguments{
\item{category_name}{names of each category. set 1 when single-species model}

\item{zone}{no. of UTM which was estimated when "Extrapolation_List" was made}

\item{labs}{setting of figure. colour is legend}

\item{ncol}{number of figures in line side by side. max is no. of "Category"}

\item{shape}{shape of COG point}

\item{size}{size of shape}

\item{package}{Spatial DeltaGLMM or FishStatsUtils}

\item{lhs, rhs}{specify what lhs and rhs are}
}
\description{
plot the COG om map using ggplot2. category names, axis names, font size and plot shape can change.
}
\examples{
# some examples if you want to highlight the usage in the package
}
